+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.2.dev] - In Development
+### Added
+- Support for ERA5 pressure levels dataset
+- Added pressure_level parameter for specifying atmospheric levels
+- Adaptive file naming for pressure level data
+
 ## [0.0.1] - 20250419
 ### Added
 - Initial version with basic ERA5 download functionality